Feature/emit events to component from dom (#14)
* feat: export the component an the function

* refactor: emit event from parent to web component to donwload the pdf

* chore: updated readme

* fix: update tests

* chore: update readme
